+++ app/about/hr.tsx
@@ -1,0 +1,3 @@
+export function HR() {
+  return <hr className="w-full bg-gray-700 my-4" />
+}
