+++ app/announcements.tsx
@@ -1,9 +1,10 @@
-import { NewsList } from '@/components/NewsList'
 import {
   getAngPararayos,
   getTheSatellite,
   getAnnouncements
 } from '@/functions/news'
+import {NewsList} from './news-list'
+import {Section} from './section'
 
 export async function AnnouncementsSection() {
   const announcements = await getAnnouncements()
@@ -11,7 +12,7 @@ export async function AnnouncementsSection() {
   const angPararayos = await getAngPararayos()
 
   return (
-    <section className="px-4 lg:px-8 py-8 lg:py-14 w-full text-white flex flex-col gap-8">
+    <Section>
       <div className="grid grid-cols-1 lg:grid-cols-3 gap-10 container">
         <div className="flex flex-col gap-4">
           <h2 className="text-2xl font-semibold">Announcements</h2>
@@ -29,6 +30,6 @@ export async function AnnouncementsSection() {
           <NewsList news={angPararayos} href="/news/ang-pararayos" />
         </div>
       </div>
-    </section>
+    </Section>
   )
 }
