+++ app/news-list.tsx
@@ -1,9 +1,14 @@
+import {News} from '@/functions/news'
 import {faArrowRight} from '@fortawesome/free-solid-svg-icons'
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome'
-import {ScrollArea} from './ui/scroll-area'
-import {NewsListProps} from './types/NewsListProps'
+import {ScrollArea} from '@/ui/scroll-area'
 import Image from 'next/image'
 import Link from 'next/link'
+
+export type NewsListProps = {
+  news: News[]
+  href: string
+}
 
 export function NewsList({news, href}: NewsListProps) {
   return (
