+++ app/news/page.tsx
@@ -1,3 +1,3 @@
 export default function News() {
-  return <></>
+  return <main></main>
 }
