+++ app/page.tsx
@@ -4,11 +4,11 @@ import {AnnouncementsSection} from './announcements'
 
 export default async function Home() {
   return (
-    <main>
+    <div>
       <HeroSection />
       <ServicesSection />
       <hr className="border-t border-neutral-700" />
       <AnnouncementsSection />
-    </main>
+    </div>
   )
 }
