+++ app/section.tsx
@@ -1,0 +1,7 @@
+export function Section({children}: {children: React.ReactNode}) {
+  return (
+    <section className="px-4 lg:px-8 py-8 lg:py-14 w-full text-white flex flex-col gap-8">
+      {children}
+    </section>
+  )
+}
