--- components/types/HeroButtonProps.ts
@@ -1,4 +1,0 @@
-export type HeroButtonProps = {
-  children: React.ReactNode
-  href: string
-}
