--- components/types/NavProps.ts
@@ -1,4 +1,0 @@
-export type NavProps = {
-  leftLinks: {name: string; href: string}[]
-  rightLinks: {name: string; href: string}[]
-}
